ci/infra/libvirt: fixed the libvirt setup
This is based on the work by Dario in #38.

First of all, I've fixed the documentation on how to fetch the image, since the
command was plain wrong. Moreover, I've also added documentation on how to
manage SSH keys, since it wasn't too clear to newcomers.

I've added the `authorized_keys` terraform variable where it was required. Plus,
I've added `depends_on` terraform instructions into the master and the worker
nodes to avoid concurrency problems: they will now wait for the LB first to
finish (we did the same in kubic-project/automation). Finally, I've changed the
names for the master/worker nodes so they don't contain underscores (which is
troublesome down the line).

Last but not least, I've also changed the cloud-init file for the load-balancer
so it starts haproxy always (before it was just enabled but it never started).

As a final note, with these changes everything seems to work except for Helm,
which is unable to forward some ports (I've seen that this is related to some
global DNS records which are expected by the provider). This is an issue to be
solved in the future.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>
Suggested-by: Dario Maiocchi <dmaiocchi@suse.de>