infra: change list of required packages
Do not require the `kubectl` package, require instead the
`kubernetes-client` one. The latter one is a sub-package of the
`kubernetes` package, while the `kubelet` one is an independent one.

The `kubernetes-client` package is **always** going to be in sync with
all the other kubernetes packages; that doesn't always stand true for
the `kubectl` one.

Having different versions of client/server packages can lead to
unexpected errors.

Finally, be explicit about the CRI we want to use: cri-o.