Add health-monitor to load-balancer
There is currently no monitor on the instances which
results in the nodes to be always seen as online in
the load-balancer pool even if the apiserver is not running.
A consequence of that is that traffic can be directed to
a node which is not able to ensure the service as it should.

The monitor will perform an HTTPS check on $MASTER:6443/healthz
an expect a 200 code. If the apiserver is not running on the master,
the node will be mark as offline so no traffic will be redirected to it.

Signed-off-by: lcavajani <lcavajani@suse.com>