Use caaspctl tooling image for node removal
Instead of using an external tooling image when performing node removal
use one that will be in our registries and that will depend on the build
type of caaspctl.

Since this change included creating an Addons concept in the image
versioning logic, refactor it to also include Cilium as well, since it's
more an addon than a control plane component.