ci: jenkins: author-check: Fix logic around git-log calls
Trying to guess the base branch for a PR is not really safe as git
is not making this information available to us. The GitHub API already
offers us the list of commit SHAs which are part of the PR so we can
get this list and examine them one by one instead of trying to obtain
a diff between a guessed base branch and the PR refspec.