ci: jenkins: check-valid-author: Ensure PR is not from local branch
It was agreed that PRs should be sent from forked repositories instead
of creating branches to the main one. We need to detect such instances
during PR testing in order to enforce this decision. This commit uses
the GitHub API to fetch the PR information and compare the head and the
base repositories. If they are the same, then further checks are skipped
and the script returns False.