ci: jenkins: handle-prs: Merge the 'clean' and 'behind' states
Now that the 'PR needs to be rebased' restriction has been removed from
GitHub, every approved PR is reported as 'clean' so we do not actually
test it against the last master. As such, we should merge the previously
reported 'behind' state with the 'clean' one to ensure that we run the
PR against the latest master before we actually merge it.