Improve CONTRIBUTING.md
Instead of adding the bugzilla IDs in body message we should take care
to include them in the commit title. We mostly use commit titles to
create the change log between releases, thus if we want to get bugfix
traces automated in OBS the bugzilla ID should be included in the
commit title.