skuba-update: drop the usage of patch-check
We were using `patch-check` before, but we realized that in some scenarios this
was not useful. Moreover, it has been noted that in some cases zypper returned 0
even if there were some patches that required the machine to reboot. These cases
have also been added in the update logic.

Finally, some refactorings have been performed in order to better accommodate
the new logic.