Enable ssh-agent in CI
This PR enables the SSH Agent during the terraform apply.
If you look at the previous CI runs, it the SSH-Agent is False.
After this PR is should be true.

This is needed because we are planning to go away from password
authentication -- which is what is currently used by the CI.