skuba-update: fixed os tests and added more tests
There were several tests which were broken for the following reasons.

First of all /etc/machine-id did not exist in containers. In order to fix that,
I've added a helper function which creates this file with a very specific
value (see later on).

Moreover, the commands which were calling `kubectl` were doing it in a wrong
way: they were passing environment variables in the command string. This will
fail with Python's `subprocess.Popen` function. Instead, we have to pass
environment variables as a separate dictionary. After fixing this, notice that
`kubectl` does not exist in our Docker images. Hence, running `skuba-update` on
os tests (SKUBA=1) would fail. In order to fix this, os tests will install a
fake kubectl script which prints what we expect and it stores which arguments
have been used so far. This will later be used with another check at the end of
each test script.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>