Better node upgrade plan output (bsc#1144914)
When control plane nodes are not yet upgraded, the skuba node upgrade
plan command run against a worker node outputs a misleading:

"Node my-worker-0 is up to date"

In this case we propagate a custom error that will enhance and
clarify the output.

Also generalize AllControlPlanesMatchVersion function.

This function can be used to check against any clusterVersion.

https://bugzilla.suse.com/show_bug.cgi?id=1144914

Signed-off-by: Maximilian Meister <mmeister@suse.de>