Add `RuntimeClass` and `StorageObjectInUseProtection` admission plugins
These two plugins are part of the recommended defaults. Their purpose:

- `RuntimeClass` [1]: It allows to choose the runtime configuration per
pod or container. It is in alpha state.

- `StorageObjectInUseProtection` [2]: It protects PV's and PVC's with a
`kubernetes.io/pv-protection` finalizer added at creation time, so
these resources won't be deleted if any pod is using them. The PVC
protection controller [3] iterates over these resources and removes
the finalizer if no pod is using them, so they are effectively only
removed if there are no pods using it.

[1] https://kubernetes.io/docs/concepts/containers/runtime-class/
[2] https://kubernetes.io/docs/concepts/storage/persistent-volumes/#storage-object-in-use-protection
[3] https://github.com/kubernetes/kubernetes/tree/master/pkg/controller/volume/pvcprotection