Do not check that node exists when upgrading it
This operation is basically a no-op, because the node already exists
on the node list when an upgrade happens, and is not waiting in any
meaningful way for the node.

Thus, remove this check from the upgrade. Also, remove this logic that
is not referenced from anywhere else. It could make sense during
bootstrap and join operations, but this check is already done by
kubeadm.