CI: VMWare make sure the var is null
Terraform has a problem with nulls and non existing vars. It will
save the vars into a json file, and of course null is not a valid
value for json, so it stores the null as a string.

Which means that in a corner case we could be checking for the null
value but instead get a "null" string which will not work for our
datastore slection based on null values.

Instead lets check for both null and "null" values.