CI: wait for sonobuoy to finish tests
Instead of running the tests and checking every 2 minutes we
can instead just launch the command and make it wait for the
test finish.

This makes our wrapper more clean by removing uneeded lines
and avoids having a complex recheck or wait when running the
tests. Timeouts are provided via jenkins so it should not
exceed the normal times in case of the command hanging