CI: Inhibit kured during conformance test
Due to the long nature of the conformance tests skuba-update might
trigger and download updates that require a reboot via kured.
This seems to affect disproportionately long running tests like
the conformance one as sooner or later kured will reboot the node.

Because the sonobuoy pod has a policy of no restart, this means
that if the pod is evicted, it wont start in another node and
the tests will fail.

This patch inhibits kured from rebooting the nodes in the
conformance pipeline by setting a manual annotation on the
kured daemonset as mentioned in the kured docs[0]

[0] https://github.com/weaveworks/kured#disabling-reboots