Provision 1.18 without hyperkube
Without this patch, 1.18 greenfields will still be provisioned with
hyperkube. This is a problem, as we decided to NOT use hyperkube
going forward.

This changes it by ensuring hyperkube conditional still exists, but
adapts based on the version to provision. When 1.17 support will
be removed, we can remove all the conditionals, and bump the
vendored libraries. Upgrades to 1.18 will just update the
configmap on the first controller to be upgraded. This could
lead to issues in mixed versions clusters (versions 1.18 and 1.18-).
However, this is supposed to be a temporary state, which sounds
acceptable to me.

Closes: https://github.com/SUSE/avant-garde/issues/1431