Ensures HyperKube flags can be removed
Without this patch, it is possible to let our code rot, and
remove the versions below 1.18 without cleaning the conditionals
or the rest of our code.

This should fix it by ensuring test failure when removing
old versions. We only need to rely on humans to understand why
the test was removed, and ensure they follow the idea through.