Use an already existing virtual network
This commit adds the possibility to make use of an already existing
virtual network for libvirt platform. If the parameter `network_name` is
provided within the terraform variables terraform will attempt to an
already existing network on libvirt that is named accordingly.
Otherwise, if no `network_name` is provided terraform will create
new terraform network that will be managed by terraform itself.