Cleanup skuba-update Makefile
The skuba-update Makefile is only used for
testing purposes, and is not used in packaging.
It most likely won't change anytime soon.

The `make install` instruction can go away:
For test purposes, we have tox (and make test).
We should probably not install things on the host
without a venv anyway with `make install` (at best
we should use `pip install -e .`).

The `make test` uses tox to install the software, and
should stay.

The other two make instructions are currently broken,
and exposed through skuba itself, so I don't expect
this to be required.