Support Patternless packages
Without this patch, skuba is not able to install the new
kind of packages (patternless).

This is a problem, as it is necessary to install
caasp5 this way, which itself is necessary because
migrations from caasp4 to caasp5 break clusters by
auto-upgrading to kubernetes 1.18 and cri-o
without using skuba.

This patch tackles both the greenfield and the upgrade
cases, where we remove the previous packages,
and install the new kind of pkgs.

Partial Fix: SUSE/avant-garde#1663