remove the patch from the package
We have a patch that removes the load balancer and CI specifics so they
are not released to our customers.

However, this patch very often needs to be rebased and is a paint to do
this. This blocks any automation we can put in place.

This is a "hacky" attempt to remove this code without having to do much
in the terraform code but instead in the packaging side.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>