Bump sonobuoy versions
Some image repos no longer work which caused the job to fail.
Updating to 0.18.5 fixes those issues for the 4.5 and 4.2 job.

Upstream change related to this:
Sonobuoy v0.18.4 now uses k8s.gcr.io as the registry to pull
the conformance image from as gcr.io/google-containers is no
longer being used and is now read-only.