Drop useless tests
Currently on skuba release there is no way of choosing the kubernetes
version. This conflicts with our current 4.2 tests in which we force
an older kubernetes version. This rises several unwanted issues which
are not real world problems so those tests are pretty much useless.

Drop the tests completely to avoid issues created by ourselves